Support organization lookup by uuid
http://www.django-rest-framework.org/api-guide/generic-views/#attributes
https://github.com/adamcharnock/django-smalluuid#django-smalluuid
