+++ lighten_api/views.py
@@ -1,0 +1,10 @@
+from rest_framework import viewsets
+
+from .models import Organization
+from .serializers import OrganizationSerializer
+
+
+class OrganizationViewSet(viewsets.ModelViewSet):
+    queryset = Organization.objects.all()
+    serializer_class = OrganizationSerializer
+    lookup_field = 'uuid'
