+++ requirements.txt
@@ -5,6 +5,7 @@ django-crispy-forms==1.6.1
 django-filter==0.15.3
 django-guardian==1.4.6
 django-rest-assured==0.2.0
+django-smalluuid==0.1.3
 djangorestframework==3.5.1
 djangorestframework-jwt==1.8.0
 factory-boy==2.7.0
@@ -22,5 +23,6 @@ PyYAML==3.12
 requests==2.11.1
 simplejson==3.9.0
 six==1.10.0
+smalluuid==0.1.4
 uritemplate==3.0.0
 whitenoise==3.2.2
